Move creating of tf/metric factory to unsafe binary
