+++ oak_functions/oak_functions_loader_unsafe/Cargo.toml
@@ -14,6 +14,9 @@ clap = { version = "*", features = ["derive"] }
 # to differentiate between public and potentially sensitive log messages.
 # TODO(#2783): Re-enable logging via the log crate once the logging approach has been redisigned.
 log = { version = "*", features = ["max_level_off", "release_max_level_off"] }
+oak_functions_extension = { path = "../extension" }
 oak_functions_loader = { path = "../loader" }
 oak_logger = { path = "../logger" }
+oak_functions_tf_inference = { path = "../experimental/tf_inference" }
+oak_functions_metrics = { path = "../experimental/metrics" }
 toml = "*"
